Add an experimental CUDA program to find the hash
Since the CPU can only do things so fast, I'd like to use GPU
acceleration to speed up the process. This simple CUDA program is able
to find hashes up to about 5 characters long, but doesn't seem to work
after that.

Eventually I'd like to integrate this with the main program instead of
compiling a separate executable.